Update HX711-multi.cpp
I was trying to use this library to measure 12 load cells with 12 hx711 modules. This library was working fine up to 4 load cells and modules, but oddly when I was trying to measure more than 5 load cells, it was not working. Then I decided to write my own code to read the signals and I found out what is wrong with this library: the HX711 datasheet clearly says that PD_SCK high time should not exceed 50 us, so as the numbers of load cells increase you pass the limit. The solution is so simple: you just have to write the results in PD_SCK low time which does not have limits.